Include LibreOfficeKit headers from today for optional use
Useful when building loolwsd on a system that doesn't have new enough
LibreOfficeKit headers available. (It makes sense to use them only if the
resulting loolwsd will then be run on a system with a new enough LibreOffice
installation anyway.)

Pass the --with-lokit-path=bundled/include configure option to use them.