Use /opt/libreoffice5.0 to match LO 5.0 beta
Note that the LibreOffice version present on the build system, if any, has no
relation to the one the packaged binaries are supposed to run against. When
building RPM or Debian packages, we use the bundled copy of the LibreOfficeKit
headers we have here, and not ones from a LibreOffice source tree.