Maintain an invalidation count for each tile
Count how many times a tile has been invalidated.
For each inalidation increase the counter and for each arrival
of the new tile decrease it.
Only cache tiles with the invalidation count equal to 0 because
when zooming, we might cancel previous requests and the tiles in the
old zoom layer are no longer valid