loleaflet: notify the server about the current part
We need this because we want to prefetch tiles from other parts too.
In order to render a tile we need to call 'setPart' and then 'paintTile'
so this might change the current part in the document.
But when the user is editing/selecting we want to work on the part that
he/she is viewing