Don't send unneccesary "tilecombine" on multiple invalidates
It usually happens that we get more "invalidateTiles" requests
from LO for one change. To avoid unnecessary tile rerenders we
remember (in transientInvalidationKeys) all tile keys which we
send to the server for rerendering and don't send the request
for the same tile. This is effective until we get back a
rerendered tile from the server (transientInvalidationKeys is
cleared).

(cherry picked from commit fdd76332db6b9c3b59864132578dc01ce19f0214)