leaflet: position cursor before image drop
So we insert the image (close to or even exactly) at the drop position,
instead of the last cursor position, which may be far away.

With this, we're closer to the desktop image drop behavior in the
browser.

(cherry picked from commit 6fb558741cec29af5cb32edc864ae0bece82e619)