loleaflet - in highlighting mode a custom cursor is used
When the user clicks on the background text color control while
editing a text document, loleaflet enter in highlighting mode (exactly
as Wrieter does).
In order to notify that loleaflet has switched to highlighting mode
the cursor changes is shape to a small bottle in the act of dropping
ink.

Change-Id: Ic9a5a636df9fd6071e38edc7e0fc2e11e037a517