loleaflet: queue messages only when we fail to connect
We should also reconsider message queueing altogether.
If we lose connection we should not have any expectation
to recover at the same state when the document is shared.
In fact it could be dangerous to send those last messages
before losing connection, lest they modify a more
recent version of the document (after others' changes).

Change-Id: I9077c1db5fe1535f1fc85436580b10c751bac2ee
Reviewed-on: https://gerrit.libreoffice.org/24348
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>
Tested-by: Ashod Nakashian <ashnakash@gmail.com>