bccu#1745: Fix Calc blinking due to popup message
Use w2overlay instead of w2popup. The former adjusts itself in
the visible screen region and fixes, and does not introduce any
additional scrollbars for the lifetime of the popup message.

The w2overlay gets removed with mouse events. So instead of
'mousedown' events originally, use 'mouseup', so that there is no
succeeding mouse events to switch the overlay off.

Also point the overlay pointer towards the button, not the label.

Change-Id: Ic6c22fecc2e32075c27b6f4647373c28d2f635ae