bccu#1583: Make scrollbar visible for Impress docs
We use 180x180 px as thumbnail size, and 2px border for selected
thumbnail on top of that which makes its 184x180.
.parts-preview needs to be expanded to 214px for
mCustomScrollbar to work properly without hiding any thumbnails
including borders. With all of this, we also need to shift
impress document offset to 214px from left to make all things fit
nicely.

Change-Id: I446918403cd43fe61f0ebfb358e6a2051014b28a