loleaflet: Fix MaxBounds for the map, bccu#1739
In case of spreadsheets, zooming in and then switching
the sheet to a new one will leave the new sheet in a situation
where panning would not be possible. This is because of incorrect
maximum bounds set on the map.

Further, this often leads to row/col offsets in spreadsheets
because document view is bound under these MaxBounds but row/col
are not.

Change-Id: I468d2fee5424e11c93a0a707fde96e687061fb89