bccu#1756: Check for empty input
In case of presentation we are getting an empty value in response
to .uno:CharFontName. While this looks like a bug in LOKit, no harm
in putting a check for empty values here.

In ES5 argument to Object.keys must be an object otherwise it
throws an exception.

Change-Id: I5c2f7963544fda55bab3f314f1b098b3c9e3bd0b
(cherry picked from commit 870b46be78a3fd09384d7e8e3e74d02cfe66614e)