Build-time configurable WSD limits
The server can now be configured at build time
to limit the total number of connections and/or
the number of open documents, at a given time.

./configure --with-max-documents=10 --with-max-connections=20
will limit the number of documents to 10 and total
number of connections (on one or all documents) to 20.

(cherry picked from commit a5314372037e0cd5187b1d4b069ca55608744273)

Change-Id: I0c73a7e906c4f567cb3da480e885524815c9cc89
Reviewed-on: https://gerrit.libreoffice.org/27204
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>
Tested-by: Ashod Nakashian <ashnakash@gmail.com>