We now get viewid parameters, too, in status: messages
So adapt the cppunit checks that test the exact number of tokens.

As such I am not sure whether it makes sense to have such checks. See
"Robustness principle". It should be enough to check that the
mandatory parameters are present, and whether there are more should be
irrelevant. Also, ideally we should accept parameters in arbitrary
order. They are named, after all.