Handle disk full situations more gracefully
Introduce new API in our Util namespace to save data to a file
safely. The data is written to a temporary file in the same directory
and after that has succeeded, it is renamed atomicaly to the intended
name. If any step of the saving fails, neither the temporay file or
the intended target (if one exists before) is left behind.

Unlike in the master branch, here we don't show any message to the
users when that happens. At least not yet. Back-porting that full
functionality was harder than expected, and I wanted to get in at
least this fix first, to avoid zero-size cached tiles.