Verify at run-time that the loolforkit program has the required capabilities
If not, log a mesage and exit.

Check the capabilities only after opening the fifo. loolwsd is waiting
in open() for loolforkit to open its end of the pipe, so if loolforkit
exits before it has opened the pipe, loolwsd will wait forever and not
notice that loolforkit has exited. Which is not ideal.

Note that there are other sanity checks done before loolforkit opens
the pipe, and if some of them fail, and loolforkit exits, loolwsd will
be hanging in the open and not notice. Need to carefully look through
that and re-factor as necessary.

Using a named pipe (FIFO) is probably pointless. We should just create
a normal pipe in loolwsd and pass the fd of the reading end on the
command-line to loolforkit.